Move main to cli/main.js
